+++ mteb/tasks/Retrieval/__init__.py
@@ -30,7 +30,6 @@ from .eng.ARCChallengeRetrieval import *
 from .eng.ArguAnaRetrieval import *
 from .eng.BrightRetrieval import *
 from .eng.ClimateFEVERRetrieval import *
-from .eng.ClimateFEVERRetrievalFast import *
 from .eng.CQADupstackAndroidRetrieval import *
 from .eng.CQADupstackEnglishRetrieval import *
 from .eng.CQADupstackGamingRetrieval import *
@@ -44,16 +43,13 @@ from .eng.CQADupstackUnixRetrieval import *
 from .eng.CQADupstackWebmastersRetrieval import *
 from .eng.CQADupstackWordpressRetrieval import *
 from .eng.DBPediaRetrieval import *
-from .eng.DBPediaRetrievalFast import *
 from .eng.FaithDialRetrieval import *
 from .eng.FeedbackQARetrieval import *
 from .eng.FEVERRetrieval import *
-from .eng.FEVERRetrievalFast import *
 from .eng.FiQA2018Retrieval import *
 from .eng.HagridRetrieval import *
 from .eng.HellaSwagRetrieval import *
 from .eng.HotpotQARetrieval import *
-from .eng.HotpotQARetrievalFast import *
 from .eng.LegalBenchConsumerContractsQARetrieval import *
 from .eng.LegalBenchCorporateLobbyingRetrieval import *
 from .eng.LegalSummarizationRetrieval import *
@@ -67,16 +63,13 @@ from .eng.LitSearchRetrieval import *
 from .eng.MedicalQARetrieval import *
 from .eng.MLQuestions import *
 from .eng.MSMARCORetrieval import *
-from .eng.MSMARCORetrievalFast import *
 from .eng.MSMARCOv2Retrieval import *
 from .eng.NarrativeQARetrieval import *
 from .eng.NFCorpusRetrieval import *
 from .eng.NQRetrieval import *
-from .eng.NQRetrievalFast import *
 from .eng.PiqaRetrieval import *
 from .eng.QuailRetrieval import *
 from .eng.QuoraRetrieval import *
-from .eng.QuoraRetrievalFast import *
 from .eng.RARbCodeRetrieval import *
 from .eng.RARbMathRetrieval import *
 from .eng.SCIDOCSRetrieval import *
@@ -91,7 +84,6 @@ from .eng.TempReasonL3ContextRetrieval import *
 from .eng.TempReasonL3FactRetrieval import *
 from .eng.TempReasonL3PureRetrieval import *
 from .eng.TopiOCQARetrieval import *
-from .eng.TopiOCQARetrievalFast import *
 from .eng.Touche2020Retrieval import *
 from .eng.TRECCOVIDRetrieval import *
 from .eng.WinoGrandeRetrieval import *
@@ -114,13 +106,10 @@ from .multilingual.CrossLingualSemanticDiscriminationWMT21 import *
 from .multilingual.IndicQARetrieval import *
 from .multilingual.MintakaRetrieval import *
 from .multilingual.MIRACLRetrieval import *
-from .multilingual.MIRACLRetrievalFast import *
 from .multilingual.MLQARetrieval import *
 from .multilingual.MultiLongDocRetrieval import *
 from .multilingual.NeuCLIR2022Retrieval import *
-from .multilingual.NeuCLIR2022RetrievalRetrievalFast import *
 from .multilingual.NeuCLIR2023Retrieval import *
-from .multilingual.NeuCLIR2023RetrievalRetrievalFast import *
 from .multilingual.PublicHealthQARetrieval import *
 from .multilingual.StatcanDialogueDatasetRetrieval import *
 from .multilingual.WikipediaRetrievalMultilingual import *
@@ -133,19 +122,14 @@ from .pol.ArguAnaPLRetrieval import *
 from .pol.DBPediaPLRetrieval import *
 from .pol.FiQAPLRetrieval import *
 from .pol.HotpotQAPLRetrieval import *
-from .pol.HotpotQAPLRetrievalFast import *
 from .pol.MSMARCOPLRetrieval import *
-from .pol.MSMARCOPLRetrievalFast import *
 from .pol.NFCorpusPLRetrieval import *
 from .pol.NQPLRetrieval import *
-from .pol.NQPLRetrievalFast import *
 from .pol.QuoraPLRetrieval import *
-from .pol.QuoraPLRetrievalFast import *
 from .pol.SCIDOCSPLRetrieval import *
 from .pol.SciFactPLRetrieval import *
 from .pol.TRECCOVIDPLRetrieval import *
 from .rus.RiaNewsRetrieval import *
-from .rus.RiaNewsRetrievalRetrievalFast import *
 from .rus.RuBQRetrieval import *
 from .slk.SlovakSumRetrieval import *
 from .spa.SpanishPassageRetrievalS2P import *
