+++ mteb/tasks/Retrieval/eng/DBPediaRetrieval.py
@@ -50,3 +50,51 @@ class DBPedia(AbsTaskRetrieval):
             },
         },
     )
+
+
+class DBPediaHardNegatives(AbsTaskRetrieval):
+    metadata = TaskMetadata(
+        name="DBPediaHardNegatives",
+        description="DBpedia-Entity is a standard test collection for entity search over the DBpedia knowledge base",
+        reference="https://github.com/iai-group/DBpedia-Entity/",
+        dataset={
+            "path": "mteb/DBPedia_test_top_250_only_w_correct",
+            "revision": "efbe63af6ea3bd1622bcd416c77b0572c089533e",
+        },
+        type="Retrieval",
+        category="s2p",
+        modalities=["text"],
+        eval_splits=["test"],
+        eval_langs=["eng-Latn"],
+        main_score="ndcg_at_10",
+        date=("2017-01-01", "2017-01-01"),  # best guess: based on publication date
+        domains=["Written", "Encyclopaedic"],
+        task_subtypes=[],
+        license="mit",
+        annotations_creators="derived",
+        dialect=[],
+        sample_creation="found",
+        bibtex_citation="""@inproceedings{Hasibi:2017:DVT,
+ author =    {Hasibi, Faegheh and Nikolaev, Fedor and Xiong, Chenyan and Balog, Krisztian and Bratsberg, Svein Erik and Kotov, Alexander and Callan, Jamie},
+ title =     {DBpedia-Entity V2: A Test Collection for Entity Search},
+ booktitle = {Proceedings of the 40th International ACM SIGIR Conference on Research and Development in Information Retrieval},
+ series =    {SIGIR '17},
+ year =      {2017},
+ pages =     {1265--1268},
+ doi =       {10.1145/3077136.3080751},
+ publisher = {ACM}
+}""",
+        descriptive_stats={
+            "n_samples": {"test": 400},
+            "avg_character_length": {
+                "test": {
+                    "average_document_length": 338.65365967056323,
+                    "average_query_length": 34.085,
+                    "num_documents": 90336,
+                    "num_queries": 400,
+                    "average_relevant_docs_per_query": 38.215,
+                }
+            },
+        },
+    )
+
