+++ mteb/tasks/Retrieval/eng/HotpotQARetrieval.py
@@ -80,3 +80,66 @@ class HotpotQA(AbsTaskRetrieval):
             },
         },
     )
+
+
+class HotpotQAHardNegatives(AbsTaskRetrieval):
+    metadata = TaskMetadata(
+        name="HotpotQAHardNegatives",
+        dataset={
+            "path": "mteb/hotpotqa_test_top_250_only_w_correct",
+            "revision": "0d4cc90462aeb15de5d11a911a3bb7bd2da0e9cb",
+        },
+        description=(
+            "HotpotQA is a question answering dataset featuring natural, multi-hop questions, with strong"
+            + " supervision for supporting facts to enable more explainable question answering systems."
+        ),
+        reference="https://hotpotqa.github.io/",
+        type="Retrieval",
+        category="s2p",
+        modalities=["text"],
+        eval_splits=["test"],
+        eval_langs=["eng-Latn"],
+        main_score="ndcg_at_10",
+        date=("2018-01-01", "2018-12-31"),  # best guess: based on publication date
+        domains=["Web", "Written"],
+        task_subtypes=["Question answering"],
+        license="cc-by-sa-4.0",
+        annotations_creators="human-annotated",
+        dialect=[],
+        sample_creation="found",
+        bibtex_citation="""@inproceedings{yang-etal-2018-hotpotqa,
+    title = "{H}otpot{QA}: A Dataset for Diverse, Explainable Multi-hop Question Answering",
+    author = "Yang, Zhilin  and
+      Qi, Peng  and
+      Zhang, Saizheng  and
+      Bengio, Yoshua  and
+      Cohen, William  and
+      Salakhutdinov, Ruslan  and
+      Manning, Christopher D.",
+    editor = "Riloff, Ellen  and
+      Chiang, David  and
+      Hockenmaier, Julia  and
+      Tsujii, Jun{'}ichi",
+    booktitle = "Proceedings of the 2018 Conference on Empirical Methods in Natural Language Processing",
+    month = oct # "-" # nov,
+    year = "2018",
+    address = "Brussels, Belgium",
+    publisher = "Association for Computational Linguistics",
+    url = "https://aclanthology.org/D18-1259",
+    doi = "10.18653/v1/D18-1259",
+    pages = "2369--2380",
+    abstract = "Existing question answering (QA) datasets fail to train QA systems to perform complex reasoning and provide explanations for answers. We introduce HotpotQA, a new dataset with 113k Wikipedia-based question-answer pairs with four key features: (1) the questions require finding and reasoning over multiple supporting documents to answer; (2) the questions are diverse and not constrained to any pre-existing knowledge bases or knowledge schemas; (3) we provide sentence-level supporting facts required for reasoning, allowing QA systems to reason with strong supervision and explain the predictions; (4) we offer a new type of factoid comparison questions to test QA systems{'} ability to extract relevant facts and perform necessary comparison. We show that HotpotQA is challenging for the latest QA systems, and the supporting facts enable models to improve performance and make explainable predictions.",
+}""",
+        descriptive_stats={
+            "n_samples": {"test": 1000},
+            "avg_character_length": {
+                "test": {
+                    "average_document_length": 357.91445598233946,
+                    "average_query_length": 92.584,
+                    "num_documents": 1043229,
+                    "num_queries": 1000,
+                    "average_relevant_docs_per_query": 2.0,
+                }
+            },
+        },
+    )
