--- mteb/tasks/Retrieval/eng/NQRetrievalFast.py
@@ -1,48 +1,0 @@
-from __future__ import annotations
-
-from mteb.abstasks.TaskMetadata import TaskMetadata
-
-from ....abstasks.AbsTaskRetrieval import AbsTaskRetrieval
-
-
-class NQFast(AbsTaskRetrieval):
-    metadata = TaskMetadata(
-        name="NQ-Fast",
-        dataset={
-            "path": "mteb/nq_test_top_250_only_w_correct",
-            "revision": "cd13dd9a902e4edbbf42a69b3cffcd8feae5daaa",
-        },
-        description="NFCorpus: A Full-Text Learning to Rank Dataset for Medical Information Retrieval",
-        reference="https://ai.google.com/research/NaturalQuestions/",
-        type="Retrieval",
-        category="s2p",
-        modalities=["text"],
-        eval_splits=["test"],
-        eval_langs=["eng-Latn"],
-        main_score="ndcg_at_10",
-        date=None,
-        domains=None,
-        task_subtypes=None,
-        license=None,
-        annotations_creators=None,
-        dialect=None,
-        sample_creation=None,
-        bibtex_citation="""@article{47761,title	= {Natural Questions: a Benchmark for Question Answering Research},
-        author	= {Tom Kwiatkowski and Jennimaria Palomaki and Olivia Redfield and Michael Collins and Ankur Parikh 
-        and Chris Alberti and Danielle Epstein and Illia Polosukhin and Matthew Kelcey and Jacob Devlin and Kenton Lee 
-        and Kristina N. Toutanova and Llion Jones and Ming-Wei Chang and Andrew Dai and Jakob Uszkoreit and Quoc Le 
-        and Slav Petrov},year	= {2019},journal	= {Transactions of the Association of Computational 
-        Linguistics}}""",
-        descriptive_stats={
-            "n_samples": {"test": 1000},
-            "avg_character_length": {
-                "test": {
-                    "average_document_length": 596.3618630215394,
-                    "average_query_length": 47.878,
-                    "num_documents": 545093,
-                    "num_queries": 1000,
-                    "average_relevant_docs_per_query": 1.2169756662804172,
-                }
-            },
-        },
-    )
