--- mteb/tasks/Retrieval/eng/QuoraRetrievalFast.py
@@ -1,54 +1,0 @@
-from __future__ import annotations
-
-from mteb.abstasks.TaskMetadata import TaskMetadata
-
-from ....abstasks.AbsTaskRetrieval import AbsTaskRetrieval
-
-
-class QuoraRetrievalFast(AbsTaskRetrieval):
-    ignore_identical_ids = True
-
-    metadata = TaskMetadata(
-        name="QuoraRetrieval-Fast",
-        dataset={
-            "path": "mteb/QuoraRetrieval_test_top_250_only_w_correct",
-            "revision": "4fd5f24639bf4e81055261ddfd1cfe44f73e4507",
-        },
-        description=(
-            "QuoraRetrieval is based on questions that are marked as duplicates on the Quora platform. Given a"
-            + " question, find other (duplicate) questions."
-        ),
-        reference="https://quoradata.quora.com/First-Quora-Dataset-Release-Question-Pairs",
-        type="Retrieval",
-        category="s2s",
-        modalities=["text"],
-        eval_splits=["test"],
-        eval_langs=["eng-Latn"],
-        main_score="ndcg_at_10",
-        date=None,
-        domains=None,
-        task_subtypes=None,
-        license=None,
-        annotations_creators=None,
-        dialect=None,
-        sample_creation=None,
-        bibtex_citation="""@misc{quora-question-pairs,
-    author = {DataCanary, hilfialkaff, Lili Jiang, Meg Risdal, Nikhil Dandekar, tomtung},
-    title = {Quora Question Pairs},
-    publisher = {Kaggle},
-    year = {2017},
-    url = {https://kaggle.com/competitions/quora-question-pairs}
-}""",
-        descriptive_stats={
-            "n_samples": {"test": 1000},
-            "avg_character_length": {
-                "test": {
-                    "average_document_length": 61.371302161846586,
-                    "average_query_length": 51.228,
-                    "num_documents": 459052,
-                    "num_queries": 1000,
-                    "average_relevant_docs_per_query": 1.5672,
-                }
-            },
-        },
-    )
