+++ mteb/tasks/Retrieval/multilingual/NeuCLIR2022Retrieval.py
@@ -128,3 +128,84 @@ class NeuCLIR2022Retrieval(MultilingualTask, AbsTaskRetrieval):
             revision=self.metadata_dict["dataset"]["revision"],
         )
         self.data_loaded = True
+
+
+
+class NeuCLIR2022RetrievalHardNegatives(MultilingualTask, AbsTaskRetrieval):
+    metadata = TaskMetadata(
+        name="NeuCLIR2022RetrievalHardNegatives",
+        description="The task involves identifying and retrieving the documents that are relevant to the queries.",
+        reference="https://neuclir.github.io/",
+        dataset={
+            "path": "mteb/neuclir-2022-fast",
+            "revision": "b94a5147a3b3046581c8b4524d9409f94594c5a4",
+            "trust_remote_code": True,
+        },
+        type="Retrieval",
+        category="s2p",
+        modalities=["text"],
+        eval_splits=["test"],
+        eval_langs=_LANGUAGES,
+        main_score="ndcg_at_20",
+        date=("2021-08-01", "2022-06-30"),
+        domains=["News", "Written"],
+        task_subtypes=[],
+        license="odc-by",
+        annotations_creators="expert-annotated",
+        dialect=[],
+        sample_creation="found",
+        bibtex_citation="""@article{lawrie2023overview,
+  title={Overview of the TREC 2022 NeuCLIR track},
+  author={Lawrie, Dawn and MacAvaney, Sean and Mayfield, James and McNamee, Paul and Oard, Douglas W and Soldaini, Luca and Yang, Eugene},
+  journal={arXiv preprint arXiv:2304.12367},
+  year={2023}
+}""",
+        descriptive_stats={
+            "n_samples": None,
+            "avg_character_length": {
+                "test": {
+                    "average_document_length": 2220.6705525948482,
+                    "average_query_length": 65.06432748538012,
+                    "num_documents": 63048,
+                    "num_queries": 342,
+                    "average_relevant_docs_per_query": 19.807142857142857,
+                    "hf_subset_descriptive_stats": {
+                        "fas": {
+                            "average_document_length": 3010.529632371083,
+                            "average_query_length": 85.4298245614035,
+                            "num_documents": 20265,
+                            "num_queries": 114,
+                            "average_relevant_docs_per_query": 14.217391304347826,
+                        },
+                        "rus": {
+                            "average_document_length": 2623.860809712729,
+                            "average_query_length": 85.58771929824562,
+                            "num_documents": 20921,
+                            "num_queries": 114,
+                            "average_relevant_docs_per_query": 24.177777777777777,
+                        },
+                        "zho": {
+                            "average_document_length": 1102.6741377733053,
+                            "average_query_length": 24.17543859649123,
+                            "num_documents": 21862,
+                            "num_queries": 114,
+                            "average_relevant_docs_per_query": 21.040816326530614,
+                        },
+                    },
+                }
+            },
+        },
+    )
+
+    def load_data(self, **kwargs):
+        if self.data_loaded:
+            return
+
+        self.corpus, self.queries, self.relevant_docs = load_neuclir_data(
+            path=self.metadata_dict["dataset"]["path"],
+            langs=self.metadata.eval_langs,
+            eval_splits=self.metadata_dict["eval_splits"],
+            cache_dir=kwargs.get("cache_dir", None),
+            revision=self.metadata_dict["dataset"]["revision"],
+        )
+        self.data_loaded = True
