--- mteb/tasks/Retrieval/pol/MSMARCOPLRetrievalFast.py
@@ -1,53 +1,0 @@
-from __future__ import annotations
-
-from mteb.abstasks.TaskMetadata import TaskMetadata
-
-from ....abstasks.AbsTaskRetrieval import AbsTaskRetrieval
-
-
-class MSMARCOPLFast(AbsTaskRetrieval):
-    ignore_identical_ids = True
-
-    metadata = TaskMetadata(
-        name="MSMARCO-PL-Fast",
-        description="MS MARCO is a collection of datasets focused on deep learning in search",
-        reference="https://microsoft.github.io/msmarco/",
-        dataset={
-            "path": "mteb/msmarco-pl_test_top_250_only_w_correct",
-            "revision": "main",
-            "trust_remote_code": True,
-        },
-        type="Retrieval",
-        category="s2p",
-        modalities=["text"],
-        eval_splits=["test"],
-        eval_langs=["pol-Latn"],
-        main_score="ndcg_at_10",
-        date=("2016-01-01", "2016-12-30"),  # best guess: based on publication date
-        domains=["Web", "Written"],
-        task_subtypes=["Question answering"],
-        license="https://microsoft.github.io/msmarco/",
-        annotations_creators="derived",
-        dialect=[],
-        sample_creation="machine-translated",
-        bibtex_citation=""""@misc{wojtasik2024beirpl,
-      title={BEIR-PL: Zero Shot Information Retrieval Benchmark for the Polish Language}, 
-      author={Konrad Wojtasik and Vadim Shishkin and Kacper Wołowiec and Arkadiusz Janz and Maciej Piasecki},
-      year={2024},
-      eprint={2305.19840},
-      archivePrefix={arXiv},
-      primaryClass={cs.IR}
-}""",
-        descriptive_stats={
-            "n_samples": {"test": 43},
-            "avg_character_length": {
-                "test": {
-                    "average_document_length": 349.3574939240471,
-                    "average_query_length": 33.02325581395349,
-                    "num_documents": 8841823,
-                    "num_queries": 43,
-                    "average_relevant_docs_per_query": 95.3953488372093,
-                }
-            },
-        },
-    )
