+++ mteb/tasks/Retrieval/pol/NQPLRetrieval.py
@@ -49,3 +49,49 @@ class NQPL(AbsTaskRetrieval):
             },
         },
     )
+
+
+class NQPLHardNegatives(AbsTaskRetrieval):
+    metadata = TaskMetadata(
+        name="NQ-PLHardNegatives",
+        description="Natural Questions: A Benchmark for Question Answering Research",
+        reference="https://ai.google.com/research/NaturalQuestions/",
+        dataset={
+            "path": "mteb/nq-pl_test_top_250_only_w_correct",
+            "revision": "main",
+            "trust_remote_code": True,
+        },
+        type="Retrieval",
+        category="s2p",
+        modalities=["text"],
+        eval_splits=["test"],
+        eval_langs=["pol-Latn"],
+        main_score="ndcg_at_10",
+        date=None,
+        domains=None,
+        task_subtypes=None,
+        license=None,
+        annotations_creators=None,
+        dialect=[],
+        sample_creation="machine-translated",
+        bibtex_citation="""@misc{wojtasik2024beirpl,
+      title={BEIR-PL: Zero Shot Information Retrieval Benchmark for the Polish Language}, 
+      author={Konrad Wojtasik and Vadim Shishkin and Kacper Wołowiec and Arkadiusz Janz and Maciej Piasecki},
+      year={2024},
+      eprint={2305.19840},
+      archivePrefix={arXiv},
+      primaryClass={cs.IR}
+}""",
+        descriptive_stats={
+            "n_samples": {"test": 1000},
+            "avg_character_length": {
+                "test": {
+                    "average_document_length": 502.14302128535564,
+                    "average_query_length": 48.31662804171495,
+                    "num_documents": 2681468,
+                    "num_queries": 3452,
+                    "average_relevant_docs_per_query": 1.2169756662804172,
+                }
+            },
+        },
+    )
